separate the discussions update worker from component
